Implement lockFile for windows
