+++ internal/multistorage/stats/cache/file.go
@@ -6,8 +6,6 @@ import (
 	"io"
 	"os"
 	"time"
-
-	"golang.org/x/sys/unix"
 )
 
 type SharedFile struct {
@@ -86,18 +84,3 @@ func (sf *SharedFile) write(content storageStatuses) error {
 	return nil
 }
 
-func lockFile(file *os.File, exclusive bool) (err error) {
-	how := unix.LOCK_SH
-	if exclusive {
-		how = unix.LOCK_EX
-	}
-
-	for {
-		err = unix.Flock(int(file.Fd()), how)
-		// When calling syscalls directly, we need to retry EINTR errors. They mean the call was interrupted by a signal.
-		if err != unix.EINTR {
-			break
-		}
-	}
-	return err
-}
